New system to the wasd
I think I already included that it will require the press of the shift along with WASD in the previous commit.. This new one uses while rather than loop and is much shorter. The delays and stops is kind of fixed, maybe I can do better idk. Also the speed is recommended to use 20-50 for best performance, the higher the number the slower it is.